requirements.txt
fastapi==0.112.0
PyJWT==2.8.0
redis==5.0.7
tortoise-orm==0.21.5
uvicorn==0.21.0
aiofiles==24.1.0
aerich